It compiles after refactoring
