+++ rudof_cli/src/input_convert_format.rs
@@ -1,9 +1,12 @@
+use crate::dctap_format::DCTapFormat as CliDCTapFormat;
+use anyhow::{bail, Result};
 use clap::ValueEnum;
-
 use std::{
     fmt::{Display, Formatter},
     str::FromStr,
 };
+
+use crate::{CliShaclFormat, ShExFormat};
 
 #[derive(Copy, Clone, PartialEq, Eq, PartialOrd, Ord, ValueEnum, Debug)]
 #[clap(rename_all = "lower")]
@@ -13,6 +16,31 @@ pub enum InputConvertFormat {
     ShExJ,
     Turtle,
     Xlsx,
+}
+
+impl InputConvertFormat {
+    pub fn to_shex_format(&self) -> Result<ShExFormat> {
+        match self {
+            InputConvertFormat::ShExC => Ok(ShExFormat::ShExC),
+            InputConvertFormat::ShExJ => Ok(ShExFormat::ShExJ),
+            InputConvertFormat::Turtle => Ok(ShExFormat::Turtle),
+            _ => bail!("Converting ShEx, format {self} not supported"),
+        }
+    }
+    pub fn to_shacl_format(&self) -> Result<CliShaclFormat> {
+        match self {
+            InputConvertFormat::Turtle => Ok(CliShaclFormat::Turtle),
+            _ => bail!("Converting to SHACL, format {self} not supported"),
+        }
+    }
+
+    pub fn to_dctap_format(&self) -> Result<CliDCTapFormat> {
+        match self {
+            InputConvertFormat::CSV => Ok(CliDCTapFormat::CSV),
+            InputConvertFormat::Xlsx => Ok(CliDCTapFormat::XLSX),
+            _ => bail!("Converting to DCTAP, format {self} not supported"),
+        }
+    }
 }
 
 impl FromStr for InputConvertFormat {
