+++ rudof_cli/src/lib.rs
@@ -1,8 +1,11 @@
 // Current modules
 pub mod cli;
+pub mod cli_shacl_format;
 pub mod color_support;
+pub mod convert;
 pub mod data;
 pub mod data_format;
+pub mod dctap;
 pub mod dctap_format;
 pub mod dctap_result_format;
 pub mod input_convert_format;
@@ -21,7 +24,9 @@ pub mod result_service_format;
 pub mod result_shacl_validation_format;
 pub mod result_shex_validation_format;
 pub mod result_validation_format;
-pub mod shacl_format;
+pub mod service;
+pub mod shacl;
+pub mod shapemap;
 pub mod shapemap_format;
 pub mod shex;
 pub mod shex_format;
@@ -29,7 +34,10 @@ pub mod show_mode;
 pub mod validation_mode;
 pub mod writer;
 
+pub use cli_shacl_format::*;
 pub use color_support::*;
+pub use convert::*;
+pub use dctap::*;
 pub use dctap_result_format::*;
 pub use input_convert_format::*;
 pub use input_convert_mode::*;
@@ -45,7 +53,9 @@ pub use result_service_format::*;
 pub use result_shacl_validation_format::*;
 pub use result_shex_validation_format::*;
 pub use result_validation_format::*;
-pub use shacl_format::*;
+pub use service::*;
+pub use shacl::*;
+pub use shapemap::*;
 pub use shapemap_format::*;
 pub use shex::*;
 pub use shex_format::*;
