+++ rudof_cli/src/mime_type.rs
@@ -1,5 +1,3 @@
-use crate::data_format::DataFormat;
-
 pub trait MimeType {
     fn mime_type(&self) -> String;
 }
