+++ rudof_cli/src/output_convert_format.rs
@@ -1,6 +1,9 @@
+use anyhow::{bail, Result};
 use clap::ValueEnum;
 
 use std::fmt::{Display, Formatter};
+
+use crate::{CliShaclFormat, ShExFormat};
 
 #[derive(Copy, Clone, PartialEq, Eq, PartialOrd, Ord, ValueEnum, Debug)]
 #[clap(rename_all = "lower")]
@@ -15,6 +18,25 @@ pub enum OutputConvertFormat {
     HTML,
     SVG,
     PNG,
+}
+
+impl OutputConvertFormat {
+    pub fn to_shex_format(&self) -> Result<ShExFormat> {
+        match self {
+            OutputConvertFormat::ShExC => Ok(ShExFormat::ShExC),
+            OutputConvertFormat::ShExJ => Ok(ShExFormat::ShExJ),
+            OutputConvertFormat::Turtle => Ok(ShExFormat::Turtle),
+            _ => bail!("Converting ShEx, format {self} not supported"),
+        }
+    }
+
+    pub fn to_shacl_format(&self) -> Result<CliShaclFormat> {
+        match self {
+            OutputConvertFormat::Default => Ok(CliShaclFormat::Internal),
+            OutputConvertFormat::Turtle => Ok(CliShaclFormat::Turtle),
+            _ => bail!("Converting to SHACL, format {self} not supported"),
+        }
+    }
 }
 
 impl Display for OutputConvertFormat {
