+++ rudof_cli/src/result_validation_format.rs
@@ -1,6 +1,8 @@
 use clap::ValueEnum;
 use std::fmt::{Display, Formatter};
 
+use crate::{ResultShExValidationFormat, ResultShaclValidationFormat};
+
 #[derive(Copy, Clone, PartialEq, Eq, PartialOrd, Ord, ValueEnum, Debug)]
 #[clap(rename_all = "lower")]
 pub enum ResultValidationFormat {
@@ -12,6 +14,34 @@ pub enum ResultValidationFormat {
     NQuads,
     Compact,
     Json,
+}
+
+impl ResultValidationFormat {
+    pub fn to_shex_result_format(&self) -> ResultShExValidationFormat {
+        match self {
+            ResultValidationFormat::Turtle => ResultShExValidationFormat::Turtle,
+            ResultValidationFormat::NTriples => ResultShExValidationFormat::NTriples,
+            ResultValidationFormat::RDFXML => ResultShExValidationFormat::RDFXML,
+            ResultValidationFormat::TriG => ResultShExValidationFormat::TriG,
+            ResultValidationFormat::N3 => ResultShExValidationFormat::N3,
+            ResultValidationFormat::NQuads => ResultShExValidationFormat::NQuads,
+            ResultValidationFormat::Compact => ResultShExValidationFormat::Compact,
+            ResultValidationFormat::Json => ResultShExValidationFormat::Json,
+        }
+    }
+
+    pub fn to_shacl_result_format(&self) -> ResultShaclValidationFormat {
+        match &self {
+            ResultValidationFormat::Turtle => ResultShaclValidationFormat::Turtle,
+            ResultValidationFormat::NTriples => ResultShaclValidationFormat::NTriples,
+            ResultValidationFormat::RDFXML => ResultShaclValidationFormat::RDFXML,
+            ResultValidationFormat::TriG => ResultShaclValidationFormat::TriG,
+            ResultValidationFormat::N3 => ResultShaclValidationFormat::N3,
+            ResultValidationFormat::NQuads => ResultShaclValidationFormat::NQuads,
+            ResultValidationFormat::Compact => ResultShaclValidationFormat::Compact,
+            ResultValidationFormat::Json => ResultShaclValidationFormat::Json,
+        }
+    }
 }
 
 impl Display for ResultValidationFormat {
